Add account create logic into a mutation
